fix: 매직넘버, 벨리데이션 메세지 추가
