+++ src/constants/index.js
@@ -9,6 +9,7 @@ import {
 	THIRD_PLACE,
 	SECOND_PLACE,
 	FIRST_PLACE,
+	MAX_WINNING_NUMBERS_LENGTH,
 } from './magicNumber.js';
 
 import {
@@ -16,11 +17,14 @@ import {
 	UNDER_MIN_NUMBER_MESSAGE,
 	OVER_MAX_NUMBER_MESSAGE,
 	DUPLICATE_NUMBER_MESSAGE,
+	INPUT_AMOUNT_MESSAGE,
+	NOT_EMPTY_WINNING_NUMBERS_MESSAGE,
 } from './validatorMessage.js';
 
 export {
 	MAX_RANDOM_NUMBER,
 	MAX_COUNT_PER_LOTTO,
+	MAX_WINNING_NUMBERS_LENGTH,
 	PRICE_PER_LOTTO,
 	FIFTH_PLACE,
 	FORTH_PLACE,
@@ -34,4 +38,6 @@ export {
 	UNDER_MIN_NUMBER_MESSAGE,
 	OVER_MAX_NUMBER_MESSAGE,
 	DUPLICATE_NUMBER_MESSAGE,
+	INPUT_AMOUNT_MESSAGE,
+	NOT_EMPTY_WINNING_NUMBERS_MESSAGE,
 };
