+++ src/js/lotto.js
