+++ src/js/rendering.js
@@ -71,4 +71,31 @@ const renderLottoResultTable = (dom, resultTable) => {
 	dom.innerHTML = tableTemplate;
 };
 
-export { renderLottoList, renderLottoResultTable, renderNumberToDOM };
+const changeDisplayBlockToNone = (element) => {
+	element.classList.remove('d-block');
+	element.classList.add('d-none');
+};
+
+const changeDisplayNoneToBlock = (element) => {
+	element.classList.remove('d-none');
+	element.classList.add('d-block');
+};
+
+const resetInputValue = (...inputsArgs) => {
+	inputsArgs.forEach((inputs) => {
+		if (inputs.length > 0) {
+			inputs.forEach((input) => (input.value = ''));
+		} else {
+			inputs.value = '';
+		}
+	});
+};
+
+export {
+	renderLottoList,
+	renderLottoResultTable,
+	renderNumberToDOM,
+	changeDisplayBlockToNone,
+	changeDisplayNoneToBlock,
+	resetInputValue,
+};
